refactor: 코드 컨벤션 적용
