+++ src/main/java/chess/domain/board/Board.java
@@ -5,7 +5,6 @@ import chess.domain.pieces.piece.Piece;
 import chess.domain.square.Movement;
 import chess.domain.square.Square;
 import chess.dto.PieceResponse;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
