+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,12 +1,17 @@
 package chess.domain.board;
 
-import chess.domain.pieces.*;
+import chess.domain.pieces.Bishop;
+import chess.domain.pieces.BlackPawn;
+import chess.domain.pieces.King;
+import chess.domain.pieces.Knight;
+import chess.domain.pieces.Queen;
+import chess.domain.pieces.Rook;
+import chess.domain.pieces.WhitePawn;
 import chess.domain.pieces.piece.Color;
 import chess.domain.pieces.piece.Piece;
 import chess.domain.square.File;
 import chess.domain.square.Rank;
 import chess.domain.square.Square;
-
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
