+++ src/main/java/chess/domain/pieces/piece/Piece.java
@@ -1,7 +1,6 @@
 package chess.domain.pieces.piece;
 
 import chess.domain.square.Movement;
-
 import java.util.Objects;
 
 public abstract class Piece {
