+++ src/main/java/chess/view/OutputView.java
@@ -1,7 +1,6 @@
 package chess.view;
 
 import chess.dto.PieceResponse;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
