+++ src/test/java/chess/domain/board/BoardTest.java
@@ -1,14 +1,15 @@
 package chess.domain.board;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import chess.domain.square.Square;
 import chess.dto.PieceResponse;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.*;
 
 @DisplayName("게임판")
 class BoardTest {
