+++ src/test/java/chess/domain/pieces/KingTest.java
@@ -1,13 +1,13 @@
 package chess.domain.pieces;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import chess.domain.pieces.piece.Color;
 import chess.domain.pieces.piece.Piece;
 import chess.domain.square.Movement;
 import chess.domain.square.Square;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("킹")
 class KingTest {
