+++ src/test/java/chess/view/PieceViewTest.java
@@ -1,10 +1,10 @@
 package chess.view;
 
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 @DisplayName("기물 출력 값")
 class PieceViewTest {
